don't allow a single id more than once in a file
this is a draconic hotfix for a painful issue:
score id is used as unique part of filename of the generated gabc file
and if it is actually not unique, only the last score with the given id
survives and appears then more then once in the preview document.

This issue should probably be addressed in a different way in the future.