'gly preview' always produces gabc files with unique suffix
problem addressed:
when a gly document contained only one score, gabc without unique suffix
(i.e. name of the document, with just the extension changed) was
created, which was then processed by gregorio producing a .tex file with
a name clashing with the preview main document name. So the main file
was overwritten.